# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A social network API built with Go. Supports user authentication, posts, comments, reactions, groups, messaging, notifications, and real-time WebSocket events.",
        "title": "Social Network API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/profile/privacy": {
            "put": {
                "description": "Sets the profile to private (1) or public (0). Private profiles require follow requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile privacy",
                "parameters": [
                    {
                        "description": "Privacy setting",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile privacy updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/update": {
            "put": {
                "description": "Updates the logged-in user's nickname, about me section, and optional avatar image.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update own profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "New nickname",
                        "name": "nickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "New about me text",
                        "name": "aboutme",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "New avatar image",
                        "name": "avatar",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/{id}": {
            "get": {
                "description": "Returns profile data for a user. For private profiles, only public info is shown unless the current user is a follower.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile/User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile data fetched",
                        "schema": {
                            "$ref": "#/definitions/repModal.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "repModal.Comment": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dislikeCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isLiked": {
                    "type": "integer"
                },
                "likeCount": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "repModal.PostResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "description": "Associated Categories",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "comment_count": {
                    "type": "integer"
                },
                "comments": {
                    "description": "Nested Comments list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repModal.Comment"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "dislike_count": {
                    "type": "integer"
                },
                "id": {
                    "description": "We embed the main Post fields directly",
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "is_liked": {
                    "description": "1: liked, 0: none, -1: disliked",
                    "type": "integer"
                },
                "like_count": {
                    "description": "Aggregated counts for the post",
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "privacy": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "time_ago": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "repModal.ProfileResponse": {
            "type": "object",
            "properties": {
                "aboutme": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repModal.UserFollow"
                    }
                },
                "following": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repModal.UserFollow"
                    }
                },
                "following_status": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_private": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repModal.PostResponse"
                    }
                }
            }
        },
        "repModal.UserFollow": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}